Currency Transfer

+++ Findaroo.Server/Model/RequestModel/User/PostUserRequest.cs
@@ -14,8 +14,8 @@
         public string occupation { get; set; }
         public string? company { get; set; }
         public string? school { get; set; }
-        public int? min_price { get; set; }
-        public int? max_price { get; set; }
+        public float? min_price { get; set; }
+        public float? max_price { get; set; }
         public int[]? lifestyle_answers { get; set; }
         public string? preferences { get; set; }
         public string? room_type { get; set; }
